core/tests/requirements.txt
# RooCode Testing Framework Requirements
# Python dependencies for comprehensive testing suite
# Version: 1.0
# Created: 2025-06-29

# Core testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# YAML processing
PyYAML>=6.0.1
ruamel.yaml>=0.17.32

# HTTP testing
requests>=2.31.0
responses>=0.23.0

# Docker testing
docker>=6.1.0

# Code quality and linting
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0

# Coverage reporting
coverage>=7.2.0

# Test data generation
faker>=19.3.0

# Async testing support
aiohttp>=3.8.0
aiofiles>=23.1.0

# Performance testing
pytest-benchmark>=4.0.0

# Parallel testing
pytest-parallel>=0.1.1

# HTML reporting
pytest-html>=3.2.0

# JSON reporting
pytest-json-report>=1.5.0

# Test discovery and organization
pytest-testmon>=2.0.0
pytest-watch>=4.2.0

# Mocking and fixtures
factory-boy>=3.3.0
freezegun>=1.2.0

# Database testing (if needed)
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0

# Configuration testing
configparser>=5.3.0
python-dotenv>=1.0.0

# Logging testing
pytest-logging>=2016.11.4

# Memory profiling
memory-profiler>=0.61.0
psutil>=5.9.0

# Type checking
mypy>=1.5.0
types-PyYAML>=6.0.0
types-requests>=2.31.0

# Documentation testing
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Security testing
bandit>=1.7.0
safety>=2.3.0

# Performance monitoring
py-spy>=0.3.0

# Test utilities
testfixtures>=7.2.0
parameterized>=0.9.0

# File system testing
pyfakefs>=5.2.0

# Time testing
time-machine>=2.10.0

# Network testing
pytest-httpserver>=1.0.0

# Environment testing
tox>=4.6.0

# Continuous integration
pytest-github-actions-annotate-failures>=0.2.0

# Test reporting
allure-pytest>=2.13.0

# API testing
tavern>=2.0.0

# Load testing
locust>=2.15.0
